refactor: split  main script into separate scripts for each pipeline step
