+++ src/modina/__init__.py
@@ -1,5 +1,9 @@
 from modina.context_simulation import simulate_copula, save_gt
-from modina.diffnet import *
+from modina.context_net_inference import compute_context_scores
+from modina.edge_filtering import filter
+from modina.diff_net_construction import compute_diff_network
+from modina.ranking import compute_ranking
+from modina.pipeline import diffnet_analysis
 
 __all__ = ["diffnet_analysis", "compute_context_scores", "filter",
            "compute_diff_network", "compute_ranking",
